Framework for new monster sounds
Implementing them will take a while again.

+++ include/permonst.h
@@ -51,7 +51,7 @@ struct permonst {
 	unsigned short	cwt,			/* weight of corpse */
 			cnutrit;		/* its nutritional value */
 	short		pxlth;			/* length of extension */
-	uchar		msound;			/* noise it makes (6 bits) */
+	int		msound;			/* noise it makes (6 bits) */
 	uchar		msize;			/* physical size (3 bits) */
         unsigned long   mresists;               /* resistances */
 	unsigned long		mconveys;		/* conveyed by eating */
